moved url, views, and services
