+++ talentmap_api/available_positions/views/available_position.py
@@ -1,6 +1,7 @@
 import coreapi
 import maya
 import logging
+import pydash
 
 from django.core.exceptions import PermissionDenied, SuspiciousOperation
 from django.shortcuts import get_object_or_404
@@ -36,6 +37,7 @@ import talentmap_api.fsbid.services.available_positions as services
 import talentmap_api.fsbid.services.projected_vacancies as pvservices
 import talentmap_api.fsbid.services.common as comservices
 import talentmap_api.fsbid.services.employee as empservices
+import talentmap_api.fsbid.services.bid as bidservices
 
 logger = logging.getLogger(__name__)
 
@@ -396,3 +398,41 @@ class AvailablePositionHighlightActionView(APIView):
         position.is_highlighted = False
         position.save()
         return Response(status=status.HTTP_204_NO_CONTENT)
+
+class BureauBiddersRankings(APIView):
+
+    permission_classes = [Or(isDjangoGroupMember('ao_user'), isDjangoGroupMember('bureau_user')), ]
+
+    schema = AutoSchema(
+        manual_fields=[
+            coreapi.Field("perdet", location='query', description='perdet of Bureau bidder'),
+        ]
+    )
+
+    def get(self, request, pk):
+        """
+        Return position information for all of bidders' bids including their ranking information for those positions
+        """
+        # 1. grab user bidlist
+        user_bids = bidservices.user_bids(pk, request.META['HTTP_JWT'])
+        # 2. grab users' rankings, excluding current position
+        user_rankings = AvailablePositionRanking.objects.filter(bidder_perdet=pk).exclude(cp_id=pk).order_by('rank')
+        # 3. filter out bids not on shortlist
+        shortlist_bids = list(filter(lambda x: (user_rankings.filter(
+            cp_id=str(pydash.get(x, 'position.id'))).exists()), user_bids))
+        filtered_bids = []
+        for bid in shortlist_bids:
+            pos_id = str(pydash.get(bid, 'position.id'))
+            hasBureauPermissions = empservices.has_bureau_permissions(pos_id, self.request.META['HTTP_JWT'])
+            hasOrgPermissions = empservices.has_org_permissions(pos_id, self.request.META['HTTP_JWT'])
+            # 4. filter out based on bureau/org perms
+            if hasOrgPermissions or hasBureauPermissions:
+                # audit how much data is being sent to FE and how much we actually need
+                bid["ranking"] = user_rankings.filter(cp_id=pos_id).values_list("rank", flat=True).first()
+                filtered_bids.append(bid)
+
+        # keep in mind that the ranking comes back as -1 the UI-Ranking
+        return Response({
+            "results": filtered_bids,
+            "other-sl-bidcount": len(shortlist_bids) - len(filtered_bids),
+        })
